requirements.txt
msal
requests
boto3
google-cloud-storage
python-dotenv
pytest